refactor(pull-request.js): simplify the error handling logic as much as possible
there are actually a number of common elements that we can reduce down to single lines. The notable
ones are: #1 The same prefix is used in every error case ("Can't submit pull request. ") #2 they all
end up using the logger.  Also, the usage of 'logger.error' will exit the process with exit code 1,
so there is no need to use a return statement to halt control flow since the whole app terminates at
that point.